requirements.txt
streamlit
langchain
langchain-community
langchain-groq
chromadb
pypdf
python-dotenv
unstructured[all-docs]
sentence-transformers